rename prelude to context, fix one more usage of old header
